WEB-1554
Extracted the search functionality from the persistence manager.  It is now all available from a search Performer.

This required a read only transaction so the enhancement of entities does not update the db.

The search actions no longer pass themselves to the search performer.  they translate themselves into an immutable version.