WEB-1631
removed logging and changed the junit libs to match the rest of the projects. from the selenium tests

Corrected the regex for the tenant name validation.  there is a selenium test making sure each type of validation fires.
there is an integration test checking the validation on the signup crud action itself.

and unit tests that check the regex does cover the action.