WEB-1701
the service locator was being used by the product merge call instead of it being injected or created.

I would appear the that the ejb container just allowed this to work.

It has been changed over to creating the manager directly and having it use the same EM as the merge call, putting everything under the same managed scope.