WEB-1708
the free form location has been moved in to the location component.

product crud sets all the values through the advanced location.   

queries and searches have been updated to target the location in the component.


column has to be renamed and uses of the get and set location method needs to be updated.


still need to add the full ui control for the predefined locations.