ApiProcedureDefinitionResource (V2) no longer needs to inherit from SetupDataResource because they're no longer all being preloaded.
ApiProcedureDefinitionResourceV2 is now its own resource, as well as a
sub-resource of ApiAssetResource in order to have clean URLs for asset
procdefs (/asset/{assetId}/procedures) and also allow for wholesale
downloading of procedures for offline assets
(/procedureDefinitionV2/list?id={ids}). Jersey doesn't like using a the
same class as a resource and a sub-resource and throws up a warning, but
the severity of that has yet to be determined.