WEB-5069 I have much less hair than when I started... but now ProcedureDefinitions can be saved through the API, instead of just read! The new method (Path: .../fieldid/api/v1/procedureDefinitionV2/save) expects an ApiProcedureDefinitionV2 object as a parameter.  This REST Service method consumes and produces application/json.
When a new ProcedureDefinition is saved, a Response code of 201 is issued, and a JSON model of the saved Entity (including associated IDs) is returned.
When an existing ProcedureDefinition is updated, a response code of 200 is issued, and a JSON model of the saved Entity (including associated IDs) is returned.
When there is a failure to save, a Response code of 500 (internal server error) is issued, and a JSON model containing only a single String describing the error is returned.