Java 8 Update
There are a number of changes here to bring us up to using Java 8:
 - Moving up to Hibernate 4.3.6 involved some small changes and is
going to require some tests to be fixed before they will stop failing
during the build (we can make those updates later)
 - S3Service was using an old Sun BASE64Encoder, which has been swapped
out for the internal Base64.Encoder.  It should work the same.
 - A small hack needed to be made due to confusion the compiler was
experiencing at compile time surrounding type erasure and class
resolution

The POMs were also updated so that we can start using Maven 3.2.2… that
means that Jenkins will probably have a heart attack when trying to run
a build… we’ll need to update that.
 - All plugins now have versions specified
 - we’re using project.version instead of pom.version
 - we’re using a lot of properties instead of keying in versions of
dependencies with literal values.  this will make updating to new
versions a lot easier, as you’ll only have to change the version
mentioned in one place.