WEB-5204 Fixed with Streams and Lambdas.
In certain situations, images and annotations would be added to a ProcedureDefinition even when the related IsolationPoint wasn't present.  This issue now can never happen.

Every time the user clicks, "Back," the contents of the entire ProcedureDefinition get checked to ensure that any Image or Annotation attached to the ProcDef is actually being used.  Those that aren't used, are removed.