Notifications are now being sent as the tenants desire: 1 email for all events that are NOT actions and 1 email for ONLY actions.
The user is only able to customize the values used for Action emails.  Non-Action emails have static subjects and subheadings in line with the old email.  Next release (2015.3.0), we should take the time to update this functionality and add customization for the Non-Action email notifications.  Adding an extra column to the action_email_customization table (and OBVIOUSLY renaming that table and associated entities) will be necessary, unless we want one table for each type of email...  This additional column should state the TYPE of email, either ACTION or NON-ACTION for now... we can further divide these at a later date.

Streams are used to separate the different types of emails, allowing us to retain use of only one driving query thus reducing the amount of pressure placed on the DB.