WEB-5761 Part of the problem with the indexer was likely that it couldn't actually FIND rows which were in an ARCHIVED state.  This is because the TenantSecurityFilter was set to NOT show archived rows.
Additional code has also been added to ensure that - no matter the outcome of processing an IndexQueueItem - each IndexQueueItem is deleted from the table.  While this could theoretically have some impact on the accuracy of indexes, those instances should be very few and far between.  Most index problems can be resolved easily by reindexing a tenant.  The important thing is to keep the size of the queue low.  The larger it grows, the longer it takes to query, lengthening the index process to the point where the table fills faster than it can be emptied.

Some cleanup was also done in AssetIndexWriter to remove an unused AutoWired Service class.