Add type filter on "latest org" query
This allows the front-end to be a little dumber with data processing, as there
is no longer a possibility of inserting an org before its parent is in the local
DB, which means we no longer need to cache a whoel ton of orgs (Intelligrated,
I'm looking at you) in memory before being able to insert.

Also, thanks Jordan!