2015.6.1-SNAPSHOT Pom update
It might LOOK like a lot of other changes are here... but there aren't.  For some reason, IntelliJ decided this was the perfect time to clean up imports in a bunch of different classes.  It wanted to see if I had a strong heart or would die from cardiac arrest.