WEB-5900 Changing the ID of the masterevents table from event_id to id should resolve the problems we were seeing retrieving ThingEvent entities from the DB.  It seems Hibernate wasn't very happy with our flip-flopping over the Primary Join column name.  This aligns us to better DB structure, anyways.
In addition to the DB script, we will altogether drop the @PrimaryKeyJoinColumn annotation, as it is only needed when the name of the PK is different between two joined tables.  We also need to change a relationship within EventAudit so that it refers to the right field.

With these changes complete. we can restore the original find(ThingEvent.class, <id>) method to the MassEventUpdateService and eliminate any risk that we would retrieve an event other than a ThingEvent.