WEB-5997 Dropping some of the "handleDownload" functionality into the base page class.
This helps with downloading generated reports or other generated files and sending them through a Link component.  Prior to this, there was some complicated work for handling simple downloads... but now we can simply call to the "handleDownload" method in the base page class to handle the generated file for us.  This call simply needs to be placed in the overridden onClick method of a Link component... it will automagically manipulate the request cycle to sneeak the file to the user.  Never underestimate this code's sneakiness.