Build stabilization.
This should make the build stable on our new development environments.

WARNING: This removes the Public WS from the builds.  We'll need to figure out how to package up the protobuf-c 1.1.1 build and protobuf 2.6.1 so that they can be loaded up on our new environments... that's the major problem.  There have been some major changes in the current version of protobuf that breaks everything.